refactor(settle-shadow): replace the fatBERA placeholder with a REAL deploy-truth consumer
fatBERA was an arbitrary DeFi example from the brief with no real data — it abstained on
placeholder coords and had nothing to do with this build. Replace it with a real,
creds-free, framework-relevant claim:

  'is the running code actually the commit we claim is deployed?'

verified against REAL local git (git rev-parse HEAD vs a pinned commit). HELD when the
deployed code matches the claim, FALSIFIED when it has drifted, INSUFFICIENT when git is
unreadable. Run against this checkout it returns FALSIFIED — the brake actually DETECTS that
HEAD has drifted from the claimed origin/main. That is the ungrounded-live-state cure doing
real work, not a token-shaped demo.

- scripts/settle-shadow/deploy-commit-truth.shadow.mjs (composes the laplas brake, SHADOW)
- 7 tests (HELD/FALSIFIED/INSUFFICIENT, never-enforces, real-git HELD)
- removes fatbera-deployed-truth.shadow.{mjs,test.mjs}

Still flips enforcement_from_below GAP→✓ (a real freeside call-site invokes the laplas gate).
